Order + start to MTGCardVariants
Changed requests to display all items in order both sets according to date and cards according to setNumber. Changed display when no image is found to show number and name instead.
Seperated parent page for card and created a new page for the eventual listing of all variants of that same card.